KimSangya/KeyboardSBA #3 아무것도 값이 없을 경우를 대비해서, 에러값 수정

+++ src/main/java/com/keyboardsba/auction/AuctionRestController.java
@@ -110,7 +110,7 @@ public class AuctionRestController {
 		
 		// 만약 가격이 낮을 경우
 		int paid = (int)Integer.parseInt(content);
-		int maxPrice = auctionBO.getItemPaidByItemId(itemId);
+		int maxPrice = (int)auctionBO.getItemPaidByItemId(itemId);
 		
 		int diff = paid - maxPrice;
